ml_predictor/requirements.txt
grpcio==1.67.1
grpcio-tools==1.67.1
pillow==10.2.0
tensorflow
python-Levenshtein
ultralytics
typing
scikit-learn
tf-keras
openpyxl
python-dotenv
opencv-python
matplotlib
numpy== 1.26.4
pandas
typer
loguru
sentence-transformers
paddlepaddle
paddleocr
Pillow
tqdm
pyyaml
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
torchvision==0.18.1+cpu
